Refactor: split Async/Controller/BusReset god files into TUs
Mechanical refactor only; no behavior changes intended.

AsyncSubsystem: split into Lifecycle/Interrupts/BusReset/CommandQueue/Diagnostics translation units.

ControllerCore: split into Lifecycle/Interrupts/Discovery/Facades translation units (ControllerCore.cpp kept as a thin stub TU).

BusResetCoordinator: split into FSM/Actions/DiscoveryDelay translation units.

Xcode: project uses file-system synchronized groups, so new .cpp files are picked up automatically.

Quality: tools/quality/tidy.sh, ./build.sh --test-only --no-bump, tools/quality/analyze.sh.
